feat: compatibility support
