+++ packages/schema/src/index.ts
@@ -1,13 +1,15 @@
 
 // Types
 import './types/global'
+
+export * from './types/compatibility'
+export * from './types/components'
 export * from './types/config'
 export * from './types/hooks'
-export * from './types/module'
-export * from './types/meta'
-export * from './types/nuxt'
-export * from './types/components'
 export * from './types/imports'
+export * from './types/meta'
+export * from './types/module'
+export * from './types/nuxt'
 export * from './types/pages'
 
 // Schema
