+++ packages/schema/src/types/compatibility.ts
@@ -1,0 +1,33 @@
+export interface NuxtCompatibility {
+  /**
+   * Required nuxt version. for example, `^2.14.0` or `>=3.0.0-27219851.6e49637`.
+   *
+   */
+  nuxt?: string
+
+  /**
+   * Special requirements for the module
+  */
+  requires?: {
+    /**
+     * Nuxt 2 bridge compatibility
+     *
+     * - `true`: Having bridge is necessary for this module
+     * - `false`: Bridge is not supported
+     * - `optional`: Module should work with/without bridge
+     */
+    bridge?: boolean | 'optional'
+  }
+}
+
+export interface NuxtCompatibilityIssue {
+  name: string
+  message: string
+}
+
+export interface NuxtCompatibilityIssues extends Array<NuxtCompatibilityIssue> {
+  /**
+   * Return formatted error message
+   */
+  toString(): string
+}
